chore: move ui_components::util code to crate-level util
