+++ src/util.rs
@@ -1,3 +1,38 @@
+use iced::Color;
+use palette::{rgb::Rgb, FromColor, Hsl};
+
+pub fn darken(color: Color, amount: f32) -> Color {
+    let mut hsl = to_hsl(color);
+
+    hsl.lightness = if hsl.lightness - amount < 0.0 {
+        0.0
+    } else {
+        hsl.lightness - amount
+    };
+
+    from_hsl(hsl)
+}
+
+pub fn lighten(color: Color, amount: f32) -> Color {
+    let mut hsl = to_hsl(color);
+
+    hsl.lightness = if hsl.lightness + amount > 1.0 {
+        1.0
+    } else {
+        hsl.lightness + amount
+    };
+
+    from_hsl(hsl)
+}
+
+fn to_hsl(color: Color) -> Hsl {
+    Hsl::from_color(Rgb::from(color))
+}
+
+fn from_hsl(hsl: Hsl) -> Color {
+    Rgb::from_color(hsl).into()
+}
+
 // TODO: This should take a `bitcoin::Amount` instead of a `u64` once we add the `bitcoin` crate as a dependency.
 pub fn format_amount_sats(amount_sats: u64) -> String {
     if amount_sats == 1 {
